Randomized CoinFlipper
The coin was flipping in a predictable way which made it easy to time clicking the button to get the results you want. Each time the flipping timer times out, it's wait_time get sets to a random duration. Also changed the names of the timers in GamePicker to me more relevant to their function.